decompress responses when the -d flag is used
